+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * `PUnsafeLiftDecl` and `PConstantDecl` instances for `PPositive`
 * `Plutarch.Internal.Builtin`, containing all types and functions that are
   essential to define anything in Plutarch
+* `evalScriptUnlimited` to `Plutarch.Evaluate` as unrestricted version of `evalScript`
 
 ## Changed